refactor(@covalent/dynamic-forms): Replace ng-packger with cli command
