+++ angular.json
@@ -200,7 +200,7 @@
     },
     "@covalent/highlight": {
       "root": "src/platform/highlight",
-      "sourceRoot": "src/platform/highlight/src",
+      "sourceRoot": "src/platform/highlight",
       "projectType": "library",
       "prefix": "lib",
       "architect": {
@@ -260,6 +260,41 @@
             "tsConfig": [
               "src/platform/core/tsconfig.lib.json",
               "src/platform/core/tsconfig.spec.json"
+            ],
+            "exclude": [
+              "**/node_modules/**"
+            ]
+          }
+        }
+      }
+    },
+    "@covalent/dynamic-forms": {
+      "root": "src/platform/dynamic-forms",
+      "sourceRoot": "src/platform/dynamic-forms",
+      "projectType": "library",
+      "prefix": "lib",
+      "architect": {
+        "build": {
+          "builder": "@angular-devkit/build-ng-packagr:build",
+          "options": {
+            "tsConfig": "src/platform/dynamic-forms/tsconfig.lib.json",
+            "project": "src/platform/dynamic-forms/ng-package.js"
+          }
+        },
+        "test": {
+          "builder": "@angular-devkit/build-angular:karma",
+          "options": {
+            "main": "src/platform/dynamic-forms/test.ts",
+            "tsConfig": "src/platform/dynamic-forms/tsconfig.spec.json",
+            "karmaConfig": "src/platform/dynamic-forms/karma.conf.js"
+          }
+        },
+        "lint": {
+          "builder": "@angular-devkit/build-angular:tslint",
+          "options": {
+            "tsConfig": [
+              "src/platform/dynamic-forms/tsconfig.lib.json",
+              "src/platform/dynamic-forms/tsconfig.spec.json"
             ],
             "exclude": [
               "**/node_modules/**"
