+++ src/platform/dynamic-forms/ng-package.js
@@ -1,3 +1,5 @@
-let ngPackageSettings = require('../ng-package-common.js');
-ngPackageSettings["dest"] = "../../../deploy/platform/dynamic-forms"
-module.exports = ngPackageSettings;
+const ngPackageSettings = require('../ng-package-common.js');
+module.exports = {
+  ...ngPackageSettings,
+  dest: '../../../deploy/platform/dynamic-forms',
+};
