+++ src/platform/dynamic-forms/tsconfig.lib.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tsconfig.app.json",
+  "compilerOptions": {
+    "baseUrl": "."
+  }
+}
