+++ src/platform/dynamic-forms/tsconfig.spec.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.spec.json",
+  "include": [
+    "**/*.spec.ts",
+    "**/*.d.ts"
+  ]
+}
