+++ src/platform/ng-package-common.js
@@ -1,5 +1,6 @@
 module.exports =
 {
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "lib": {
     "entryFile": "index.ts",
     "umdModuleIds": {
